requirements.txt
flask
pdfplumber
pymysql
boto3
python-dotenv
